requirements.txt
# Core dependencies
streamlit==1.32.2
python-dotenv==1.0.1

# Audio and video processing
yt-dlp==2025.3.27
pydub==0.25.1
pytube==15.0.0

# Google services
google-cloud-speech==2.31.1
google-generativeai==0.8.4
googletrans-py==4.0.0

# Text-to-speech
gTTS==2.5.0

# Vector embeddings
sentence-transformers==2.5.1
faiss-cpu==1.10.0 